If filetype is unknown, also put '#' in delim map
If NERD Commenter does not have a delim map for a filetype, it parses
the &commentstring variable. This is fine when Vim knows the filetype
and has a suitable comment string for it, but for unknown filetypes Vim
defaults to '/*%s*/', and that is far less common than files with '# %s'
comments.

Solution: if &filetype is empty, parse &commentstring is into
the alt delims, and set '#' as the main left delim.